requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncer==0.0.8
attrs==25.3.0
autocrop==1.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
dlib==19.24.8
exceptiongroup==1.2.2
ExifRead==3.0.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.3.2
gradio==5.25.2
gradio_client==1.8.0
groovy==0.1.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jsonpickle==3.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
onnxruntime-gpu==1.21.1
opencv-python-headless==4.11.0.86
orjson==3.10.16
packaging==25.0
pandas==2.2.3
perseus-core-library==1.20.5
photoidmagick==1.1.6
pillow==10.4.0
pipfile==0.0.2
platformdirs==4.3.7
pooch==1.8.2
propcache==0.3.1
protobuf==6.30.2
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
PyMatting==1.1.13
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
rembg==2.0.65
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
ruff==0.11.6
safehttpx==0.1.6
scikit-image==0.25.2
scipy==1.15.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.3
tifffile==2025.3.30
toml==0.10.2
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
Unidecode==1.3.8
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
websockets==15.0.1
yarl==1.20.0
